notebooks/sam3-uv-test/pyproject.toml
[project]
name = "sam3-uv-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "marimo>=0.19.4",
    "requests>=2.32.5",
    "segment-anything",
    "supervision>=0.27.0",
    "timm>=1.0.24",
    "ultralytics-opencv-headless>=8.4.7",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
